SOLID: added a cast in dxf_solid_write().

+++ src/solid.c
@@ -683,7 +683,7 @@ dxf_solid_write
         }
         if (solid->paperspace == DXF_PAPERSPACE)
         {
-                fprintf (fp->fp, " 67\n%d\n", DXF_PAPERSPACE);
+                fprintf (fp->fp, " 67\n%d\n", (int16_t) DXF_PAPERSPACE);
         }
         fprintf (fp->fp, "  8\n%s\n", solid->layer);
         if (strcmp (solid->linetype, DXF_DEFAULT_LINETYPE) != 0)
